Update class has been added for saying what values have been edited.
In update function in DatabaseConnector if the primary key is 0 then it is skipped due to the primary key should never be this value and it is the default value for ints.

In SqliteDatabaseLink in Execute and ExecuteReader there are replace functions to get rid of bytes as they are used in MYSQL and SQLSERVER but SQLite uses strings for everything and the byte characteristic for booleans doesn't exist.